Add UI changes to allow liking and un-liking and to show like count.
This is where it gets interesting with Turbo. I've not worked with it
before and it seems really cool - although in this case it caused a
clash as I wanted to use devise ('current_user') in my views to determine
if a user can see the 'Like' button. To get around this, I've had to
move the broadcasting event from the Post model to the PostController so
I can pass through 'current_user' as a local variable.

This might well end up causing some other issues with Turbo, and longer
term I'd want to find a solution without using those Devise variables.